requirements.txt
torch>=2.0.0
pillow>=9.0.0
diffusers @ git+https://github.com/huggingface/diffusers
transformers>=4.48.0
accelerate>=0.20.0
safetensors>=0.3.0
huggingface_hub>=0.20.0
peft>=0.13.0


